use is ready poll in from url and base uploads
